+++ src/gui/views/__init__.py
@@ -3,7 +3,5 @@ from .news import show_news
 from .privacy import privacy_statement
 from .contact import contact_form, send_contact_form
 from .profile import edit_profile
-from .courses import show_courses, course_details, course_lesson
-from django.shortcuts import render
-from django.http import HttpResponse
-
+from .courses import show_courses, course_details
+from .lessons import lesson_details, get_root_lesson_ids, get_lessons
